requirements.txt
certifi==2021.10.8
click==8.0.2
Flask==2.0.2
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.2
kafka-python==2.0.2
MarkupSafe==2.0.1
minio==7.1.1
psycopg2-binary==2.9.1
six==1.16.0
SQLAlchemy==1.4.25
urllib3==1.26.7
Werkzeug==2.0.2
